requirements.txt
apache-airflow-providers-amazon
pandas
beautifulsoup4
requests
sqlalchemy
psycopg2-binary
pytest
